+++ packages/near-api-ts/universal/src/client/methods/transaction/getTransactionResult/handleRpcResult/_common/getProcessingSteps/getNonConversionSteps/executionSteps/deserializeExecutionSteps.ts
@@ -8,15 +8,15 @@ import type {
 import type { MaybeBaseDeserializeTransactionExecutionStepsFn } from '../../../../../../../../../../types/_common/transactionDetails/transactionResult';
 import type { InnerGetTransactionResultArgs } from '../../../../../../../../../../types/client/methods/transaction/getTransactionResult';
 import { type NatError, resultNatError } from '../../../../../../../../../_common/natError';
+import { tryBase64ToObject } from '../../../../../../../../../_common/utils/base64ToObject';
 import { result } from '../../../../../../../../../_common/utils/result';
-import { tryParseBase64ToObject } from '../../../../../../../../../_common/utils/tryParseBase64ToObject';
 import { baseGetActionSummary } from '../../_common/getActionSummaries';
 
 const getParsedResult = (
   rawResult: ExecutionStepResult<Base64String>,
 ): ExecutionStepResult<unknown> =>
   rawResult.status === 'Success'
-    ? { status: 'Success', data: tryParseBase64ToObject(rawResult.data) }
+    ? { status: 'Success', data: tryBase64ToObject(rawResult.data) }
     : rawResult;
 
 const getParsedExecutionStep = (rawExecutionStep: RawExecutionStep): ParsedExecutionStep => ({