+++ packages/react-near-ts/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [UNRELEASED] v0.4.0
+
+### Changed
+- Migrated to near-api-ts v0.11.0
+
 ## v0.3.0
 
 ### Changed